Added models and calls

+++ internal/database/models/inventoryReceipt.go
@@ -5,7 +5,7 @@ import "github.com/uptrace/bun"
 type InventoryReceipt struct {
 	bun.BaseModel `bun:"table:inventory_receipts_hdr"`
 
-	ReceiptNumber float64 `bun:"receipt_number,pk"`
+	ReceiptNumber int32 `bun:"receipt_number,pk"`
 
 	InventoryReceiptItems []InventoryReceiptLine `bun:"rel:has-many,join:receipt_number=receipt_number"`
 }
